Feat: 코스 삭제 구현

+++ src/components/schedule/CourseItem.jsx
@@ -2,28 +2,35 @@ import React, { useState } from "react";
 import { api } from "../../api/api";
 import KeyboardDoubleArrowDownIcon from "@mui/icons-material/KeyboardDoubleArrowDown";
 import DeleteOutlineIcon from "@mui/icons-material/DeleteOutline";
+import sortBy from "lodash/sortBy";
 
-const CourseItem = ({ dayCourse, currentDay }) => {
-  const deleteCourse = async (id, category) => {
-    try {
-      const { data } = await api.delete(`/auth/course/details`, {
-        category,
-        detailId: id
-      });
-      if (data.isSuccess) {
-        alert("코스가 삭제되었습니다.");
-      } else {
-        alert(data.message);
+const CourseItem = ({ dayCourse, setDayCourse, currentDay }) => {
+  const deleteCourse = async (detailId, category) => {
+    if (confirm("코스를 삭제할까요?")) {
+      try {
+        const { data } = await api.delete(`/auth/course/details`, {
+          data: {
+            category,
+            detailId
+          }
+        });
+        const newDayCourse = dayCourse.filter(dayCourse => dayCourse.detailId !== detailId);
+        if (data.isSuccess) {
+          setDayCourse(newDayCourse);
+        } else {
+          alert(data.message);
+        }
+      } catch (error) {
+        throw new Error(error);
       }
-    } catch (error) {
-      throw new Error(error);
     }
+    return;
   };
 
   if (dayCourse.length === 0) return <div className="mb-6 text-center font-bold text-base">{currentDay}일차에 등록된 코스가 없습니다.</div>;
   return (
     <div>
-      {dayCourse.map(course => (
+      {sortBy(dayCourse, ["detailOrder"]).map(course => (
         <div key={course.detailOrder}>
           <div className="flex items-center flex justify-between md:mb-3 mb-1">
             <div className="flex items-center">
@@ -38,7 +45,7 @@ const CourseItem = ({ dayCourse, currentDay }) => {
             <DeleteOutlineIcon
               className="cursor-pointer"
               onClick={() => {
-                deleteCourse(course.id, course.category);
+                deleteCourse(course.detailId, course.category);
               }}
             />
           </div>
